+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 - Support for configuring claims supported in Keycloak OP metadata [RCIAM-899](https://jira.argo.grnet.gr/browse/RCIAM-899)
 ### Fixed
 - Include 'urn:ietf:params:oauth:grant-type:token-exchange' in grant_types_supported field of Keycloak OP metadata, if token-exchange is enabled [RCIAM-915](https://jira.argo.grnet.gr/browse/RCIAM-915)
+- Improve Keycloak SAML IdP SAML message bindings [RCIAM-942](https://jira.argo.grnet.gr/browse/RCIAM-942)
 
 ## [v16.1.0-rc1.0.7] - 2022-03-17
 ### Changed